LayoutTests/fast/ruby/ruby-justification-hittest-expected.txt

This test makes sure that hit testing works with ruby inside text-align: justify.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS right is 792
PASS document.elementFromPoint(right - 1, bottom - 1).id is "ruby"
PASS successfullyParsed is true

TEST COMPLETE
abcdefg abcdefga mmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmm
